+++ app/Http/Controllers/AppUser/FavoriteController.php
@@ -20,56 +20,55 @@ class FavoriteController extends Controller
         return response()->json($favorites);
     }
 
-   
+
     public function store(Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'user_id' => ['required'],
             'apartment_id' => ['required'],
         ]);
 
         if ($validator->fails()) {
-            return response()->json(['success' => "false", 'error' => 'Empty fields'], 200);
+            return response()->json(['success' => "false", 'error' => 'Empty fields'], 422);
         }
-        Favorit::create([
-            // 'user_id' => Auth::guard('app_users')->user()->id,
-            'user_id' => $request->user_id,
+        $favorit =  Favorit::create([
+            'user_id' => Auth::guard('app_users')->user()->id,
             'apartment_id' => $request->apartment_id,
-            
+
         ]);
-        return response()->json(['isSuccess' => true], 200);
+        return response()->json(['isSuccess' => true,'data' =>   $favorit], 200);
 
     }
 
-    
+
     public function show(string $id)
     {
         //
     }
 
-        
-    public function update(Request $request, Favorit $favorit)
+
+    public function update(Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'user_id' => ['required'],
             'apartment_id' => ['required'],
         ]);
-    
+
         if ($validator->fails()) {
-            return response()->json(['success' => false, 'error' => 'Empty fields'], 200);
+            return response()->json(['success' => false, 'error' => 'Empty fields'],422);
         }
-    
-        $favorit->update([
-            'user_id' => $request->user_id,
-            'apartment_id' => $request->apartment_id,
-        ]);
-    
+
+        $favorit =  favorit::where('user_id', Auth::guard('app_users')->user()->id)->first();
+        if( $favorit){
+            $favorit->delete();
+        }else{
+            Favorit::create([
+                'user_id' => Auth::guard('app_users')->user()->id,
+                'apartment_id' => $request->apartment_id,
+
+            ]);
+        }
+
         return response()->json(['isSuccess' => true], 200);
     }
 
-    public function destroy(Favorit $favorit)
-    {
-        $favorit->delete();
-        return response()->json(['isSuccess' => true], 200);
-    }
+
 }
